+++ scripts/winesapos-install.sh
@@ -539,6 +539,9 @@ fi
 # Configure higher polling frequencies for better compatibility with input devices.
 sed -i s'/GRUB_CMDLINE_LINUX="/GRUB_CMDLINE_LINUX="usbhid.jspoll=1 usbhid.kbpoll=1 usbhid.mousepoll=1 /'g /mnt/etc/default/grub
 
+# Configure the "none" I/O scheduler for better performance on flash and SSD devices.
+sed -i s'/GRUB_CMDLINE_LINUX="/GRUB_CMDLINE_LINUX="elevator=none /'g /mnt/etc/default/grub
+
 arch-chroot /mnt grub-mkconfig -o /boot/grub/grub.cfg
 echo "Setting up the bootloader complete."